chrome/test/data/xr/e2e_test_files/html/test_local_floor_reference_space_rejects.html

<!doctype html>
<!--
Tests that a request for a stationary reference space is rejected.
-->
<html>
  <head>
    <link rel="stylesheet" type="text/css" href="../resources/webxr_e2e.css">
  </head>
  <body>
    <script src="../../../../../../third_party/blink/web_tests/resources/testharness.js"></script>
    <script src="../resources/webxr_e2e.js"></script>
    <script>
      let session_ = null;
      console.log('Requesting inline session');
      navigator.xr.requestSession('inline')
      .then((xrSession) => {
        session_ = xrSession;
        console.log(
            'Got inline session, requesting local-floor reference space');
        return xrSession.requestReferenceSpace('local-floor');
      })
      .then((refSpace) => {
        assert_unreached("Should not be able to get a stationary reference space");
      }, (err) => {
        assert_not_equals(session_, null, "Ensure session was created");
        assert_equals(err.name, "NotSupportedError", "Ensure we got the right error");
        done();
      });
    </script>
  </body>
</html>
